+++ index.html
@@ -3,7 +3,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
   <title>AI Sandbox Game</title>
-  <link rel="stylesheet" href="styles.css">
+  <link rel="stylesheet" href="assets/styles/styles.css">
   <script type="importmap">
     {
       "imports": {
@@ -71,7 +71,7 @@
   
   <!-- Journal Interface -->
   <div id="journal-button">
-    <img src="Journal.png" alt="Journal" />
+    <img src="assets/images/Journal.png" alt="Journal" />
   </div>
   
   <!-- Journal Sidebar -->
@@ -107,6 +107,6 @@
     <div id="close-chat">✕</div>
   </div>
   
-  <script src="app.js" type="module"></script>
+  <script src="src/app.js" type="module"></script>
 </body>
 </html>
