+++ src/npc/NPCManager.js
