+++ src/npc/npcBehavior.js
