+++ src/npc/npcInteraction.js
