+++ src/npc/npcRenderer.js
