+++ src/player/controls.js
