+++ src/player/player.js
