+++ src/ui/characterGenerator.js
