+++ src/ui/ui.js
@@ -1,4 +1,4 @@
-import { getScreenPosition } from './screenUtils.js';
+import { getScreenPosition } from '../utils/screenUtils.js';
 
 export class UIManager {
   constructor(gameManager, npcManager) {
